Moved argument-handling functions into their own file
